SELECTION SCREEN PLS READ
Ok so selection screen works for the host but not the client bc the
client has to be spawed in spawnNewHost but it doesn't know what the
client selected bc they connect after server has been made. So if you
connect the client right now, it is always an assault and the host is
whatever you choose.